Complete Refactor of code
Right now, I don't have a device to test this on.  I'm hopeful that it
will work… but more than likely there will be some more changes first.
The workflow right now is:

1. Starts in ShakeawayActivity.java (I thought we might as well rename
while doing the refactor).
2. When "Start" button is pushed, creates GameActivity and passes in
the robot's unique ID.  We should render the scoring page here.
3. GameActivity creates the Game, which has two Team objects.  When we
call game.startGame(), the game's logic is executed.  The main part is
done inside a single while loop, with one iteration of the loop per
"turn."
4. When the game is over, I think GameActivity should finish
automatically.  I'm not sure if we will wait for the "Play Again"
button, or just immediately return to ShakeawayActivity.java.